pyproject.toml
[project]
name = "agentic-phidata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=7.2.0",
    "groq>=0.13.1",
    "phidata>=2.7.6",
    "yfinance>=0.2.51",
]
